Update patch set 1
Patch Set 1:

I agree that the apt rules are not useful here when they are out of date.  That's why I'm working to make them updatable in Debian and Ubuntu.  For example, here is a pending update of the udev rules in Debian/stable:
https://bugs.debian.org/935308

Next up is getting this update into Ubuntu/LTS. So from my point of view, this effort is definitely not abandoned.

Patch-set: 1