Update patch set 1
Patch Set 1:

This is a thought, a HOWTO or perhaps FAQ for `repo`.  `repo` is a bit of an odd tool some notes about it really need to be kept /somewhere/.  Some initial topics I think are worthy of mention are present in this change.

Figuring out how to work with multiple branches is an issue.  When interacting with two device maintainers I learned both of them simply have two entirely separate areas for the two branches they often work with.  Certainly this works, but seems very suboptimal.

Cleaning older cruft out of a given branch is an issue.

More things which need mention?

Anyone up for filling in appropriate instructions for the appropriate headings?

Patch-set: 1
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>