Added warning that installation instructions for payton brick device.
The installation instructions for payton don't take into acocunt that the
Moto X4 uses an A/B partioning scheme. I've added a warning that they
shouldn't be followed, a pointer to alternative instructions, and a pointer
to a howto for unbricking devices.

See <https://gitlab.com/LineageOS/issues/infra/issues/62#note_260384075>
for details.

Change-Id: I738b2895a11cb2e7e0e7782392fb25fe3ecbd477