Update recovery_install_fastboot_xiaomi.md
The explanation was horribly out of date. The cooldown period described in the article appears to be no longer necessary, as Xiaomi appears to have changed their policies. At least I was able to unlock the bootloader of my Pocophone F1 today without any waiting involved. If this waiting is still necessary for other Xiaomi devices, the articles need to be split.