Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I don't see any use in adding even more data. As stated previously, our wiki is not a "find your next device" go-to, so plenty of the data doesn't need to be there as it is already known to the people who own the phones
> headphone jack, ip rating, ram type, notch, mass, screen body ratio, sim size...
> also, I don't know why you don't reuse width, height, length fields and introduce new ones.

If the wiki is for people who already own the phones, why include any device information at all? A person with the device can just get all they need from that.

Also, RAM and Storage types were already included in some of the devices. I made new width, height, depth, ram, and storage fields as these new ones are stored as numbers instead of strings, making it easier for a computer to work with and make the wiki as a whole more consistent in formatting. This is not about turning the wiki itself into a place to search for devices. I believe there is value in having the information about LineageOS devices in a central location that is easy for anyone to work with for whatever they may want to do. Sure, maybe screen body ratio or notch aren't necessary to that goal and I can remove those. 

My changes also organize the data in a better way, outside of adding new information.

I apologize if I've struck a nerve, I'm just trying to help make the resources I use better.

Patch-set: 1
Reviewer: Gerrit User 24558 <24558@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>