Update patch set 1
Patch Set 1:

> Are you okay with this?

- I don't like what you are doing with bluetooth - we have a clear separation between spec and profiles already. If something is missing, extend, don't replace https://github.com/LineageOS/lineage_wiki/blob/master/test/schema-06.yml#L167-L201
- Fingerprint: can stay a peripheral
- Headphone jack: Can be added to valid peripherals list
- Comments (like "mass in grams") probably won't help, as people will most likely keep them, making the data different across devices again. If something is unclear from the name or from the example itself, we have a page explaining various properties already: https://wiki.lineageos.org/addingdevice-howto.html#populating-the-yaml
- All dimension fields: I like the conversion outside of the yaml, so keep the names and remove the values in inches
- storage: this field still waits for expansion to support something like
model1: 32GB,
model3, model4: 64GB
can keep unit there. don't need storage type
- sim: Dual/Triple/whatever can stay/become a peripheral, sim size doesn't matter
- ram: don't need type, also can stay in the current format, don't see a benefit in removing the unit - we don't do calculations
- notch: don't see a reason to add this info
- screen_body_ratio: don't need
camera: can separate between front and rear, but don't really think we need the additional info  like focus. Could add something like "features: HDR, ..." as a list, like with bluetooth profile or something
- usb: not sure why, but ok.

The main part on the page displaying all the device info is to provide a small overview of the device and by all means, mainly bring the user to the target page: how to build, how to install, etc

For the two-step approach: that's less reviewable. Convert the sample and the page in one go, get agreement on what you did, then add all the device conversions to the same patch.
It's harder to review and to no benefit (for us) to have that separated or backwards compatible. Even with this being two patches, none would go in without the other.
Before anything goes in, the validation schema needs to be adjusted as well (this patch)

Keep in mind that the device folder has 289 devices and I have often enough changed all of them, so make sure that you really want to change every field or if some can't stay as is and you just fiddle with how that data is displayed

Patch-set: 1
Reviewer: Gerrit User 13028 <13028@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>