Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> Looking good. Wonder if there's a clean way of deduplicating this? Perhaps splitting some of the device build page sections into individual templares. Just seems a bit of a pain to keep both this and the "build *device*" templates up to date.

I've thought about that (as already mentioned in Slack, but not sure how actively you are reading there), but I don't feel like we should pick apart the instructions that much (unless you really want to go for it and properly split up all the sections), at least for the time where we are actively iterating on it. 

Considering how many minor changes there are, we'd have to do quite a lot of ifdefing anyways (or duplicating those sections only, if we decide to split stuff up) if we want to go for unified templates.

The only part that I can imagine being worth it to split up/unify for is the pre-build setup (i.e. everything from installing platform-tools up to initializing the source repository, and maybe ccache if we move that further upwards), because that is the only thing I could see that would be emulator-adaptable with a few simple defines.

Patch-set: 2
Reviewer: Gerrit User 15734 <15734@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>