wiki: Correct button combinations on espressowifi
This was verified on a GT-P5113TS, where the download and recovery mode
combinations were swapped compared to the instructions. I suspect this
has something to do with the volume keys being swapped compared to how
you'd expect them to be on a phone (i.e. the Volume Down button is
closer to the power button, higher up when held in portrait mode). The
device's download mode instructions confirmed that Volume Up + Power is
the correct combination.

Change-Id: Ic86dca22aa0efc11c91d653d236650ed51bd4aae