Update patch set 8
Patch Set 8:

(2 comments)

Yeah, this is a lot better.

The changes in "What you'll need" and "Downloading the source code" would also have to be applied to the other device build pages, since the text originally came from there.

Would it be reasonable to split up the page a bit more and put everything after "Enter the following to initialize the repository" up to (but not including) "Start the build" into a seperate file?
That way, we could reference that file from both build pages, similar to how it's done for the build host setup already.

I'm not sure if splitting out the "What you'll need" section is worth it. Maybe we can add the whole section to the top of build_host_setup and add some ifs to not show a device name/to skip ahead to "Starting the build" instead of "Downloading device-specific files"?

Patch-set: 8
Reviewer: Gerrit User 15734 <15734@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>