Update patch set 6
Patch Set 6: Code-Review-2

Just putting a hold on this since it was discovered and discussed a couple weeks ago in slack that this has some unintended changes.   One being, the link font weight was greatly reduced on some pages.
But I do agree in the need for this change, as the links on some pages are hard to spot.

Patch-set: 6
Reviewer: Gerrit User 7118 <7118@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2