Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> We have a whole page "Devices" which show current and past devices and which can be filteres by that. I don't think we need to link to the build targets

I'm aware of the "Devices" page. In June I was wondering if my device (m8) is still supported. The device page said so. But there hasn't been any update published for a month, while the dual sim version (m8d) still got updates.
The FAQ says this could be a temporary issue or the device has been removed from the build roster and points to the device page. With this information I wanted to figure out if the support has been dropped and this has not been updated in the wiki.

It took me a while to figure out where to check this. At the end I found the information I was looking for. So it's fine for me now.
But I want to make this easier for other users.
* The build roster is mentioned in the FAQ.
* The lineage-build-targets file has not been renamed since January 2017 (link will be stable).
* The link can help users to help themselves (device is still there or has been removed, check reason of removal, ...).

For me the question is: Why not to add this link?

Patch-set: 1