Update patch set 1
Patch Set 1:

> Patch Set 1:
> I'm aware of the "Devices" page. In June I was wondering if my device (m8) is still supported. The device page said so. But there hasn't been any update published for a month, while the dual sim version (m8d) still got updates.
> The FAQ says this could be a temporary issue or the device has been removed from the build roster and points to the device page. With this information I wanted to figure out if the support has been dropped and this has not been updated in the wiki.
> 
> For me the question is: Why not to add this link?

I don't see any value in directing users to that file. You have all the info there, as a user. If a build is missing that is either
- when its discontinued - you'll see that on the wiki page already (and as you said, the FAQ already points to the device page for that case) or
- when there are, as we had it, issues with the build servers
For the first case this change doesn't add value, for the second you won't see that on that file either

> For me the question is: Why not to add this link?
For me, the use of the info there is questionable or not there at all. For your own case it didn't, right?

Patch-set: 1
Reviewer: Gerrit User 13028 <13028@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>