device_upgrade: Update adb root instructions
Clarify that adb root is only needed to boot into sideload on 16.0
and below.
The wiki would ask the user to select the root access option when
upgrading from 17.1 to 18.1 on devices that support 16.0, 17.1 and
18.1 but on since 17.1 that setting is not present anymore and
furthermore adb root is not required for adb reboot sideload, so
clarify that this step is only needed for 16.0 and below.

Remove enabling developer settings for devices that don't support
16.0 or below.
This is only required to enable adb root, hence move it to the same
condition.

Remove unnecessary ' at the end of the adb root instruction.

Change-Id: I8d967ab1477d7084807d41f0d06a4ad9a6b5ea08