wiki: jfltespr is not a Verizon device
We mistakenly stated jfltespr as a Verizon device in the Wiki, when
this device was never sold on Verizon. Instead, this build was
designed for the Sprint and US Cellular versions of the Samsung
Galaxy S4.

Unlike newer devices, the Sprint/US Cellular Galaxy S4 is unable to
operate on Verizon and vice versa.

Change-Id: I133766642069deadbeeb13376664206901031994