rs988: Change "2G CDMA" to "3G CDMA2000"
This is to be consistent with the formatting on the other device pages.

Although most CDMA networks have now shut down, the majority (if not all)
current and prior CDMA networks were upgraded to CDMA2000.

CDMA2000 being backwards compatible means a rs988 operating on CdmaOne (2G
CDMA) is unlikely, and therefore will almost always operate in CDMA2000
mode should a GSM, UMTS, or LTE network be unavailable on the carrier used.

Change-Id: I2666050782315415240374004093999944069428