Add instruction how to avoid duplicate `Change-Id:` in commit messages
The pre-commit hook uses `git interpret-trailers --parse` to extract
all trailers and checks if there already is a `Change-Id` in them.
If not one is generated and appended.

This fails when inbetween the `Bug:`, `Test:`, `Tag:`, `Merged-In:`
trailer lines used by AOSP a cherry-pick line appears, i.e.:
> (cherry picked from https://<URL>/q/commit:<SHA>)

Reason is that none of those trailers are known to git and due to the
added line there is no single block consisting of lines looking like
trailers, i.e. `key: value`
We can add `Change-Id` as a "user-configured trailer" which makes git
consider this block as trailers.

Change-Id: I7553369057ebabe498c11a2f00d3198a9e73cb8b