Update patch set 1
Patch Set 1: Code-Review-2

hey, thanks for the contribution, but this is not the correct way to do this either.

the correct way is just to pass a lineage zip to extract-files, which it now support globally.

we just haven't updated this document yet but I will.

the -2 stands for the cheetah change too, as it's also not the proper way to do s.

Patch-set: 1
Reviewer: Gerrit User 16185 <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2, 395f7b67a54d2d1ed06e16d582914bd91bcdfc81