betalm: use the fastboot button combo for download_boot
Unlike most recent LG devices, betalm has a fastboot.
download_boot is used for the installation instructions when using any
fastboot method. Since we're using that for this device, the button
combination displayed in the installation instructions led to LG's
download mode instead of fastboot. Switch to the fastboot button
combination to fix this.

Change-Id: Ie6088af0a7e4925c0cc20d115024d2f18e4b3b7e