joan: replace bootloader requirement with firmware requirement for fastboot variant
Previously, as the remaining joan variants were not listed on the wiki,
we enforced a specific bootloader known to support fastboot.

This is no longer the case, with non-fastboot-and-non-TMO devices being
split into another variant page, so drop this requirement.

It didn't work anyway, as `ro.bootloader` doesn't exist on LG.

Change-Id: Iea2ae59e63ebc10381ef7866a7ee1dcef7818c4b