wiki: Update information post-migration to Play Integrity
SafetyNet was officially turned down on January 31, 2025 [1], therefore
it can be expected that users will face issues with Play Integrity
instead of SafetyNet.

However, retain the link to overview of SafetyNet page on android.com,
as the migration may cause confusion to users. A "deprecated" text is
included as reference, with deprecation timeline omitted as it is shown
prominently with the overview.

Additionally, include the blog post in the Quirks section, so as to
provide more resolution for the reader.

[1]: https://developer.android.com/privacy-and-security/safetynet/deprecation-timeline

Signed-off-by: Ricky Cheung <rcheung844@gmail.com>
Change-Id: Id27fa557b9bdc4fcb7870b0b0c3edd4858271e56